Combine functionality is impelemnted
Splitting the deep zoom script into a generic command line utility, a
core library, and the implementation.

Combine works as far as I can tell. The command line prompt for it is
not correct currently and needs cleaned up. It operates entirely
synchronously right now.

Fixed a bug with maxlevel calculation. Google claimed log(2) was ~0.3…
and I used that as the constant creating zoom level 38 when it should
be 17 for the test data set.